Move pointer, touch and keyboard out of WaylandState
The seat's three devices are one concern: they arrive together with the
capabilities, they are released together when the seat loses them, and touch
is folded into the pointer pipeline rather than being its own thing. They
move to `platform::input` as `InputState`, with the seat, pointer, touch and
keyboard handlers, the cursor-shape call, the keysym mapping and the scroll
constant.

`latest_input_serial` goes with them, which is what makes the boundary
honest: it is produced by input and only borrowed by the selections, so
`selections.rs` now reads `self.input.latest_input_serial` instead of
sharing a field at the top level. The popup grab reads it the same way.

`SeatHandler` moves whole even though it also binds the selection devices —
a seat capability is an input event, and the one line it owes the clipboard
is a call to `Selections::attach_devices`.

wayland.rs 2211 -> 1498 lines, 35 -> 21 fields.
Since the start of this branch: 2665 -> 1498, 50 -> 21.
