+++ src/platform/mod.rs
@@ -1,3 +1,4 @@
+pub mod input;
 pub mod selections;
 pub mod wayland;
