+++ src/platform/wayland.rs
@@ -2,29 +2,27 @@ use raw_window_handle::{
     DisplayHandle, HandleError, HasDisplayHandle, HasWindowHandle, RawDisplayHandle,
     RawWindowHandle, WaylandDisplayHandle, WaylandWindowHandle, WindowHandle,
 };
+use smithay_client_toolkit::reexports::calloop::LoopHandle;
+use smithay_client_toolkit::reexports::client::{
+    Connection, Dispatch, EventQueue, Proxy, QueueHandle, WEnum, delegate_noop,
+    globals::registry_queue_init,
+    protocol::{wl_output, wl_surface},
+};
+use smithay_client_toolkit::reexports::protocols::xdg::shell::client::xdg_positioner;
 use smithay_client_toolkit::{
     compositor::{CompositorHandler, CompositorState, Region},
     data_device_manager::DataDeviceManagerState,
-    primary_selection::PrimarySelectionManagerState,
-    delegate_compositor, delegate_keyboard, delegate_layer, delegate_output,
-    delegate_pointer, delegate_registry, delegate_seat,
+    delegate_compositor, delegate_layer, delegate_output, delegate_registry, delegate_session_lock,
+    delegate_xdg_popup,
     output::{OutputHandler, OutputState},
+    primary_selection::PrimarySelectionManagerState,
     registry::{ProvidesRegistryState, RegistryState},
     registry_handlers,
-    seat::{
-        keyboard::{KeyEvent, KeyboardHandler, Keysym, Modifiers as WlModifiers, RawModifiers},
-        pointer::{
-            cursor_shape::CursorShapeManager, PointerEvent, PointerEventKind, PointerHandler,
-        },
-        Capability, SeatHandler, SeatState,
-        touch::TouchHandler,
-    },
-    delegate_session_lock, delegate_touch,
+    seat::{SeatState, pointer::cursor_shape::CursorShapeManager},
     session_lock::{
         SessionLock, SessionLockHandler, SessionLockState, SessionLockSurface,
         SessionLockSurfaceConfigure,
     },
-    delegate_xdg_popup,
     shell::{
         WaylandSurface,
         wlr_layer::{
@@ -37,34 +35,22 @@ use smithay_client_toolkit::{
         },
     },
 };
-use smithay_client_toolkit::reexports::calloop::LoopHandle;
-use smithay_client_toolkit::reexports::protocols::xdg::shell::client::xdg_positioner;
-use smithay_client_toolkit::reexports::client::{
-    globals::registry_queue_init,
-    protocol::{
-        wl_keyboard, wl_output, wl_pointer, wl_seat, wl_surface,
-        wl_touch,
-    },
-    Connection, Dispatch, EventQueue, Proxy, QueueHandle, WEnum, delegate_noop,
-};
-use smithay_client_toolkit::reexports::protocols::wp::cursor_shape::v1::client::wp_cursor_shape_device_v1::Shape as WpCursorShape;
 use wayland_backend::sys::client::ObjectId;
 use wayland_protocols::ext::background_effect::v1::client::{
-    ext_background_effect_manager_v1::{self, Capability as BgCapability, ExtBackgroundEffectManagerV1},
+    ext_background_effect_manager_v1::{
+        self, Capability as BgCapability, ExtBackgroundEffectManagerV1,
+    },
     ext_background_effect_surface_v1::ExtBackgroundEffectSurfaceV1,
 };
 
 use std::collections::HashMap;
 
+use super::input::InputState;
 use super::selections::Selections;
 use crate::blur::BlurRect;
 use crate::outputs::{self, OutputId, OutputInfo};
-use crate::reactive::CursorIcon;
 use crate::surface::SurfaceId;
-use crate::widgets::{Event, Key, Modifiers, MouseButton, Rect, ScrollSource};
-
-/// Pixels per line for discrete scroll (mouse wheel)
-const SCROLL_PIXELS_PER_LINE: f32 = 40.0;
+use crate::widgets::{Event, Rect};
 
 /// The shell role of a surface: an ordinary layer-shell surface or an
 /// ext-session-lock-v1 lock surface. Both share the same widget tree, GPU
@@ -209,41 +195,11 @@ pub struct WaylandState {
     /// Lock lifecycle events, drained by the main loop.
     lock_events: Vec<LockEvent>,
 
-    // Pointer state
-    pointer: Option<wl_pointer::WlPointer>,
-    pointer_x: f32,
-    pointer_y: f32,
-    pointer_over_surface: bool,
-    pointer_enter_serial: u32,
-
-    // Touch state
-    touch: Option<wl_touch::WlTouch>,
-    /// Fingers currently down: id → (surface, x, y).
-    touch_fingers: HashMap<i32, (SurfaceId, f32, f32)>,
-    /// The finger driving pointer emulation (the first one down). Widgets
-    /// only understand pointer events, so the primary finger synthesizes
-    /// MouseMove/MouseDown/MouseUp — a tap becomes a click.
-    primary_finger: Option<i32>,
-
-    // Cursor shape
-    cursor_shape_manager: Option<CursorShapeManager>,
-
-    // Keyboard state
-    keyboard: Option<wl_keyboard::WlKeyboard>,
-    modifiers: Modifiers,
-    pub(super) keyboard_serial: u32,
-    /// Track raw_code → Key for press/release matching (handles compose sequences)
-    pressed_keys: HashMap<u32, Key>,
-    /// Key repeat runs on a calloop timer owned by the toolkit, armed with the
-    /// rate and delay the compositor reports. The keyboard is created inside a
-    /// seat capability callback, which has no other route to the loop.
-    loop_handle: LoopHandle<'static, WaylandState>,
+    /// Pointer, touch and keyboard — see [`super::input`].
+    pub(super) input: InputState,
 
     /// Clipboard and primary selection — see [`super::selections`].
     pub(super) selections: Selections,
-
-    /// Serial of the most recent input event, needed to claim selections.
-    pub(super) latest_input_serial: u32,
 }
 
 /// Why the platform layer could not start or continue.
@@ -369,22 +325,8 @@ pub fn create_wayland_app(
         session_lock_state,
         active_lock: None,
         lock_events: Vec::new(),
-        pointer: None,
-        pointer_x: 0.0,
-        pointer_y: 0.0,
-        pointer_over_surface: false,
-        pointer_enter_serial: 0,
-        touch: None,
-        touch_fingers: HashMap::new(),
-        primary_finger: None,
-        cursor_shape_manager,
-        keyboard: None,
-        modifiers: Modifiers::default(),
-        keyboard_serial: 0,
-        pressed_keys: HashMap::new(),
-        loop_handle,
+        input: InputState::new(cursor_shape_manager, loop_handle),
         selections: Selections::new(data_device_manager, primary_selection_manager),
-        latest_input_serial: 0,
     };
 
     Ok((connection, event_queue, state, qh))
@@ -647,7 +589,9 @@ impl WaylandState {
         if config.grab
             && let Some(seat) = self.seat_state.seats().next()
         {
-            popup.xdg_popup().grab(&seat, self.latest_input_serial);
+            popup
+                .xdg_popup()
+                .grab(&seat, self.input.latest_input_serial);
         }
 
         wl_surface.commit();
@@ -1107,44 +1051,6 @@ impl WaylandState {
             .values()
             .any(|s| !s.first_frame_presented || !s.scale_factor_received)
     }
-
-    /// Set the cursor shape
-    pub fn set_cursor(&self, cursor: CursorIcon, qh: &QueueHandle<Self>) {
-        let Some(ref manager) = self.cursor_shape_manager else {
-            return;
-        };
-        let Some(ref pointer) = self.pointer else {
-            return;
-        };
-
-        // Convert our CursorIcon to Wayland cursor shape
-        let shape = match cursor {
-            CursorIcon::Default => WpCursorShape::Default,
-            CursorIcon::Text => WpCursorShape::Text,
-            CursorIcon::Pointer => WpCursorShape::Pointer,
-            CursorIcon::Crosshair => WpCursorShape::Crosshair,
-            CursorIcon::Move => WpCursorShape::Move,
-            CursorIcon::NotAllowed => WpCursorShape::NotAllowed,
-            CursorIcon::Grab => WpCursorShape::Grab,
-            CursorIcon::Grabbing => WpCursorShape::Grabbing,
-            CursorIcon::ResizeNorth => WpCursorShape::NResize,
-            CursorIcon::ResizeSouth => WpCursorShape::SResize,
-            CursorIcon::ResizeEast => WpCursorShape::EResize,
-            CursorIcon::ResizeWest => WpCursorShape::WResize,
-            CursorIcon::ResizeNorthEast => WpCursorShape::NeResize,
-            CursorIcon::ResizeNorthWest => WpCursorShape::NwResize,
-            CursorIcon::ResizeSouthEast => WpCursorShape::SeResize,
-            CursorIcon::ResizeSouthWest => WpCursorShape::SwResize,
-            CursorIcon::ColResize => WpCursorShape::ColResize,
-            CursorIcon::RowResize => WpCursorShape::RowResize,
-            CursorIcon::Wait => WpCursorShape::Wait,
-            CursorIcon::Progress => WpCursorShape::Progress,
-        };
-
-        // Get cursor shape device and set shape
-        let device = manager.get_shape_device(pointer, qh);
-        device.set_shape(self.pointer_enter_serial, shape);
-    }
 }
 
 pub struct WaylandWindowWrapper {
@@ -1378,395 +1284,6 @@ impl LayerShellHandler for WaylandState {
     }
 }
 
-impl SeatHandler for WaylandState {
-    fn seat_state(&mut self) -> &mut SeatState {
-        &mut self.seat_state
-    }
-
-    fn new_seat(&mut self, _conn: &Connection, _qh: &QueueHandle<Self>, _seat: wl_seat::WlSeat) {}
-
-    fn new_capability(
-        &mut self,
-        _conn: &Connection,
-        qh: &QueueHandle<Self>,
-        seat: wl_seat::WlSeat,
-        capability: Capability,
-    ) {
-        // Handle pointer capability
-        if capability == Capability::Pointer && self.pointer.is_none() {
-            log::info!("Pointer capability available, creating pointer");
-            match self.seat_state.get_pointer(qh, &seat) {
-                Ok(pointer) => self.pointer = Some(pointer),
-                Err(e) => {
-                    // A capability race at seat init is not fatal — the app
-                    // just runs without pointer input until the seat updates
-                    log::warn!("Failed to get pointer: {e}");
-                    return;
-                }
-            }
-        }
-
-        // Handle touch capability
-        if capability == Capability::Touch && self.touch.is_none() {
-            log::info!("Touch capability available, creating touch");
-            match self.seat_state.get_touch(qh, &seat) {
-                Ok(touch) => self.touch = Some(touch),
-                Err(e) => {
-                    log::warn!("Failed to get touch: {e}");
-                }
-            }
-        }
-
-        // Handle keyboard capability
-        if capability == Capability::Keyboard && self.keyboard.is_none() {
-            log::info!("Keyboard capability available, creating keyboard");
-            let loop_handle = self.loop_handle.clone();
-            let keyboard = match self.seat_state.get_keyboard_with_repeat(
-                qh,
-                &seat,
-                None,
-                loop_handle,
-                Box::new(|state: &mut WaylandState, _kbd, event| state.emit_key_repeat(event)),
-            ) {
-                Ok(keyboard) => keyboard,
-                Err(e) => {
-                    log::warn!("Failed to get keyboard: {e}");
-                    return;
-                }
-            };
-            self.keyboard = Some(keyboard);
-
-            // Selections hang off the seat, so they cannot be created with
-            // their managers at startup.
-            self.selections.attach_devices(qh, &seat);
-        }
-    }
-
-    fn remove_capability(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _seat: wl_seat::WlSeat,
-        capability: Capability,
-    ) {
-        if capability == Capability::Pointer {
-            log::info!("Pointer capability removed");
-            if let Some(pointer) = self.pointer.take() {
-                pointer.release();
-            }
-        }
-        if capability == Capability::Keyboard {
-            log::info!("Keyboard capability removed");
-            if let Some(keyboard) = self.keyboard.take() {
-                keyboard.release();
-            }
-        }
-        if capability == Capability::Touch {
-            log::info!("Touch capability removed");
-            if let Some(touch) = self.touch.take() {
-                touch.release();
-            }
-            self.touch_fingers.clear();
-            self.primary_finger = None;
-        }
-    }
-
-    fn remove_seat(&mut self, _conn: &Connection, _qh: &QueueHandle<Self>, _seat: wl_seat::WlSeat) {
-    }
-}
-
-impl TouchHandler for WaylandState {
-    fn down(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _touch: &wl_touch::WlTouch,
-        serial: u32,
-        _time: u32,
-        surface: wl_surface::WlSurface,
-        id: i32,
-        position: (f64, f64),
-    ) {
-        self.latest_input_serial = serial;
-        let Some(surface_id) = self.surface_lookup.get(&surface.id()).copied() else {
-            return;
-        };
-        let (x, y) = (position.0 as f32, position.1 as f32);
-        self.touch_fingers.insert(id, (surface_id, x, y));
-
-        // The first finger down drives pointer emulation: move + press so
-        // hover and pressed state layers respond, and a tap becomes a click.
-        if self.primary_finger.is_none() {
-            self.primary_finger = Some(id);
-            if let Some(surface_state) = self.surfaces.get_mut(&surface_id) {
-                surface_state.pending_events.push(Event::MouseMove { x, y });
-                surface_state.pending_events.push(Event::MouseDown {
-                    x,
-                    y,
-                    button: MouseButton::Left,
-                });
-            }
-        }
-    }
-
-    fn up(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _touch: &wl_touch::WlTouch,
-        _serial: u32,
-        _time: u32,
-        id: i32,
-    ) {
-        let Some((surface_id, x, y)) = self.touch_fingers.remove(&id) else {
-            return;
-        };
-        if self.primary_finger == Some(id) {
-            self.primary_finger = None;
-            if let Some(surface_state) = self.surfaces.get_mut(&surface_id) {
-                surface_state.pending_events.push(Event::MouseUp {
-                    x,
-                    y,
-                    button: MouseButton::Left,
-                });
-                // Unlike a real pointer, nothing hovers after lifting the
-                // finger — clear hover state.
-                surface_state.pending_events.push(Event::MouseLeave);
-            }
-        }
-    }
-
-    fn motion(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _touch: &wl_touch::WlTouch,
-        _time: u32,
-        id: i32,
-        position: (f64, f64),
-    ) {
-        let Some(finger) = self.touch_fingers.get_mut(&id) else {
-            return;
-        };
-        let (x, y) = (position.0 as f32, position.1 as f32);
-        finger.1 = x;
-        finger.2 = y;
-        let surface_id = finger.0;
-
-        if self.primary_finger == Some(id)
-            && let Some(surface_state) = self.surfaces.get_mut(&surface_id)
-        {
-            // Coalesce runs of MouseMove like the pointer path does.
-            let events = &mut surface_state.pending_events;
-            if let Some(last @ Event::MouseMove { .. }) = events.last_mut() {
-                *last = Event::MouseMove { x, y };
-            } else {
-                events.push(Event::MouseMove { x, y });
-            }
-        }
-    }
-
-    fn shape(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _touch: &wl_touch::WlTouch,
-        _id: i32,
-        _major: f64,
-        _minor: f64,
-    ) {
-    }
-
-    fn orientation(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _touch: &wl_touch::WlTouch,
-        _id: i32,
-        _orientation: f64,
-    ) {
-    }
-
-    fn cancel(&mut self, _conn: &Connection, _qh: &QueueHandle<Self>, _touch: &wl_touch::WlTouch) {
-        // The compositor took over the gesture: release the synthesized
-        // press and clear hover so no widget is stuck pressed.
-        if let Some(id) = self.primary_finger.take()
-            && let Some((surface_id, x, y)) = self.touch_fingers.get(&id).copied()
-            && let Some(surface_state) = self.surfaces.get_mut(&surface_id)
-        {
-            surface_state.pending_events.push(Event::MouseUp {
-                x,
-                y,
-                button: MouseButton::Left,
-            });
-            surface_state.pending_events.push(Event::MouseLeave);
-        }
-        self.touch_fingers.clear();
-    }
-}
-
-impl PointerHandler for WaylandState {
-    fn pointer_frame(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _pointer: &wl_pointer::WlPointer,
-        events: &[PointerEvent],
-    ) {
-        for event in events {
-            // Try to find the surface ID for this event's wl_surface
-            let surface_id = self.surface_lookup.get(&event.surface.id()).copied();
-
-            // Get the target event queue for this surface
-            let target_events: Option<&mut Vec<Event>> = if let Some(id) = surface_id {
-                self.surfaces.get_mut(&id).map(|s| &mut s.pending_events)
-            } else if !matches!(event.kind, PointerEventKind::Leave { .. }) {
-                // Not our surface and not a leave event, skip
-                continue;
-            } else {
-                None
-            };
-
-            match event.kind {
-                PointerEventKind::Enter { serial } => {
-                    self.pointer_over_surface = true;
-                    self.pointer_enter_serial = serial;
-                    self.latest_input_serial = serial;
-                    self.pointer_x = event.position.0 as f32;
-                    self.pointer_y = event.position.1 as f32;
-
-                    // Track which surface has pointer focus
-                    self.current_pointer_surface = surface_id;
-
-                    if let Some(events) = target_events {
-                        events.push(Event::MouseEnter {
-                            x: self.pointer_x,
-                            y: self.pointer_y,
-                        });
-                        events.push(Event::MouseMove {
-                            x: self.pointer_x,
-                            y: self.pointer_y,
-                        });
-                    }
-                }
-                PointerEventKind::Leave { .. } => {
-                    if self.pointer_over_surface {
-                        self.pointer_over_surface = false;
-
-                        // Send leave event to the surface that had focus
-                        if let Some(id) = self.current_pointer_surface
-                            && let Some(surface_state) = self.surfaces.get_mut(&id)
-                        {
-                            surface_state.pending_events.push(Event::MouseLeave);
-                        }
-
-                        self.current_pointer_surface = None;
-                    }
-                }
-                PointerEventKind::Motion { .. } => {
-                    self.pointer_x = event.position.0 as f32;
-                    self.pointer_y = event.position.1 as f32;
-                    if let Some(events) = target_events {
-                        // Coalesce runs of MouseMove: only the latest position
-                        // matters for hover state, and every queued move costs
-                        // a full event-dispatch walk of the widget tree.
-                        if let Some(last @ Event::MouseMove { .. }) = events.last_mut() {
-                            *last = Event::MouseMove {
-                                x: self.pointer_x,
-                                y: self.pointer_y,
-                            };
-                        } else {
-                            events.push(Event::MouseMove {
-                                x: self.pointer_x,
-                                y: self.pointer_y,
-                            });
-                        }
-                    }
-                }
-                PointerEventKind::Press { button, serial, .. } => {
-                    self.latest_input_serial = serial;
-                    if let Some(mouse_button) = wayland_button_to_mouse_button(button)
-                        && let Some(events) = target_events
-                    {
-                        events.push(Event::MouseDown {
-                            x: self.pointer_x,
-                            y: self.pointer_y,
-                            button: mouse_button,
-                        });
-                    }
-                }
-                PointerEventKind::Release { button, .. } => {
-                    if let Some(mouse_button) = wayland_button_to_mouse_button(button)
-                        && let Some(events) = target_events
-                    {
-                        events.push(Event::MouseUp {
-                            x: self.pointer_x,
-                            y: self.pointer_y,
-                            button: mouse_button,
-                        });
-                    }
-                }
-                PointerEventKind::Axis {
-                    horizontal,
-                    vertical,
-                    source,
-                    ..
-                } => {
-                    // Determine scroll source
-                    let scroll_source = match source {
-                        Some(wl_pointer::AxisSource::Wheel) => ScrollSource::Wheel,
-                        Some(wl_pointer::AxisSource::Finger) => ScrollSource::Finger,
-                        Some(wl_pointer::AxisSource::Continuous) => ScrollSource::Continuous,
-                        Some(wl_pointer::AxisSource::WheelTilt) => ScrollSource::Wheel,
-                        _ => ScrollSource::Wheel,
-                    };
-
-                    // Calculate delta in pixels.
-                    // Wheel steps by preference: value120 (wl_pointer v8+,
-                    // fractional steps from high-resolution wheels) then
-                    // legacy discrete. Touchpad/finger scroll has neither
-                    // and uses absolute (already in pixels).
-                    let wheel_steps =
-                        |scroll: &smithay_client_toolkit::seat::pointer::AxisScroll| {
-                            if scroll.value120 != 0 {
-                                scroll.value120 as f32 / 120.0
-                            } else {
-                                scroll.discrete as f32
-                            }
-                        };
-
-                    let steps_x = wheel_steps(&horizontal);
-                    let delta_x = if steps_x != 0.0 {
-                        steps_x * SCROLL_PIXELS_PER_LINE
-                    } else {
-                        horizontal.absolute as f32
-                    };
-
-                    let steps_y = wheel_steps(&vertical);
-                    let delta_y = if steps_y != 0.0 {
-                        steps_y * SCROLL_PIXELS_PER_LINE
-                    } else {
-                        vertical.absolute as f32
-                    };
-
-                    // Only emit scroll event if there's actual scroll delta
-                    if (delta_x != 0.0 || delta_y != 0.0)
-                        && let Some(events) = target_events
-                    {
-                        events.push(Event::Scroll {
-                            x: self.pointer_x,
-                            y: self.pointer_y,
-                            delta_x,
-                            delta_y,
-                            source: scroll_source,
-                        });
-                    }
-                }
-            }
-        }
-    }
-}
-
 /// Map guido's popup anchor point to the xdg_positioner anchor.
 fn popup_point_to_anchor(point: crate::surface::PopupAnchor) -> xdg_positioner::Anchor {
     use crate::surface::PopupAnchor as P;
@@ -1797,232 +1314,6 @@ fn popup_point_to_gravity(point: crate::surface::PopupAnchor) -> xdg_positioner:
         P::TopRight => xdg_positioner::Gravity::TopRight,
         P::BottomRight => xdg_positioner::Gravity::BottomRight,
     }
-}
-
-/// Convert Wayland button code to MouseButton
-fn wayland_button_to_mouse_button(button: u32) -> Option<MouseButton> {
-    // Linux input event codes (from linux/input-event-codes.h)
-    const BTN_LEFT: u32 = 0x110;
-    const BTN_RIGHT: u32 = 0x111;
-    const BTN_MIDDLE: u32 = 0x112;
-
-    match button {
-        BTN_LEFT => Some(MouseButton::Left),
-        BTN_RIGHT => Some(MouseButton::Right),
-        BTN_MIDDLE => Some(MouseButton::Middle),
-        _ => None,
-    }
-}
-
-impl KeyboardHandler for WaylandState {
-    fn enter(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        surface: &wl_surface::WlSurface,
-        _serial: u32,
-        _raw: &[u32],
-        _keysyms: &[Keysym],
-    ) {
-        log::debug!("Keyboard focus entered");
-
-        // Track which surface has keyboard focus
-        let surface_id = self.surface_lookup.get(&surface.id()).copied();
-        self.current_keyboard_surface = surface_id;
-
-        // Route event to correct surface
-        if let Some(id) = surface_id
-            && let Some(surface_state) = self.surfaces.get_mut(&id)
-        {
-            surface_state.pending_events.push(Event::FocusIn);
-        }
-    }
-
-    fn leave(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        surface: &wl_surface::WlSurface,
-        _serial: u32,
-    ) {
-        log::debug!("Keyboard focus left");
-
-        // Route event to correct surface
-        let surface_id = self.surface_lookup.get(&surface.id()).copied();
-        if let Some(id) = surface_id
-            && let Some(surface_state) = self.surfaces.get_mut(&id)
-        {
-            surface_state.pending_events.push(Event::FocusOut);
-        }
-
-        self.current_keyboard_surface = None;
-    }
-
-    fn press_key(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        serial: u32,
-        event: KeyEvent,
-    ) {
-        // Track serial for clipboard operations
-        self.keyboard_serial = serial;
-        self.latest_input_serial = serial;
-
-        if let Some(key) = keysym_to_key(event.keysym, event.utf8.as_deref(), true) {
-            // Store raw_code → Key mapping so release_key can emit the correct Key
-            // (e.g., composed 'é' instead of raw 'e' after a compose sequence)
-            self.pressed_keys.insert(event.raw_code, key);
-
-            let key_event = Event::KeyDown {
-                key,
-                modifiers: self.modifiers,
-            };
-
-            // Route to the surface with keyboard focus
-            if let Some(id) = self.current_keyboard_surface
-                && let Some(surface_state) = self.surfaces.get_mut(&id)
-            {
-                surface_state.pending_events.push(key_event);
-            }
-        }
-    }
-
-    fn release_key(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        _serial: u32,
-        event: KeyEvent,
-    ) {
-        // Use the stored key from press_key if available (handles compose sequences
-        // where the composed character differs from the raw keysym on release)
-        let key = self
-            .pressed_keys
-            .remove(&event.raw_code)
-            .or_else(|| keysym_to_key(event.keysym, event.utf8.as_deref(), false));
-
-        if let Some(key) = key {
-            let key_event = Event::KeyUp {
-                key,
-                modifiers: self.modifiers,
-            };
-
-            // Route to the surface with keyboard focus
-            if let Some(id) = self.current_keyboard_surface
-                && let Some(surface_state) = self.surfaces.get_mut(&id)
-            {
-                surface_state.pending_events.push(key_event);
-            }
-        }
-    }
-
-    fn update_modifiers(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        _serial: u32,
-        modifiers: WlModifiers,
-        _raw_modifiers: RawModifiers,
-        _layout: u32,
-    ) {
-        self.modifiers = Modifiers {
-            ctrl: modifiers.ctrl,
-            alt: modifiers.alt,
-            shift: modifiers.shift,
-            logo: modifiers.logo,
-        };
-    }
-
-    fn repeat_key(
-        &mut self,
-        _conn: &Connection,
-        _qh: &QueueHandle<Self>,
-        _keyboard: &wl_keyboard::WlKeyboard,
-        _serial: u32,
-        event: KeyEvent,
-    ) {
-        self.emit_key_repeat(event);
-    }
-}
-
-impl WaylandState {
-    /// Deliver a repeated key as an ordinary press.
-    ///
-    /// Two sources reach here: the toolkit's calloop timer, armed from the
-    /// compositor's `repeat_info`, and the protocol's own repeated key state.
-    /// Neither is visible to widgets — a held key looks exactly like someone
-    /// pressing it very fast.
-    fn emit_key_repeat(&mut self, event: KeyEvent) {
-        let Some(key) = keysym_to_key(event.keysym, event.utf8.as_deref(), true) else {
-            return;
-        };
-        let key_event = Event::KeyDown {
-            key,
-            modifiers: self.modifiers,
-        };
-
-        // Route to the surface with keyboard focus
-        if let Some(id) = self.current_keyboard_surface
-            && let Some(surface_state) = self.surfaces.get_mut(&id)
-        {
-            surface_state.pending_events.push(key_event);
-        }
-    }
-}
-
-/// Convert XKB keysym to our Key type
-fn keysym_to_key(keysym: Keysym, utf8: Option<&str>, is_press: bool) -> Option<Key> {
-    // Named keys first
-    match keysym {
-        Keysym::BackSpace => return Some(Key::Backspace),
-        Keysym::Delete => return Some(Key::Delete),
-        Keysym::Return | Keysym::KP_Enter => return Some(Key::Enter),
-        Keysym::Tab | Keysym::ISO_Left_Tab => return Some(Key::Tab),
-        Keysym::Escape => return Some(Key::Escape),
-        Keysym::Left => return Some(Key::Left),
-        Keysym::Right => return Some(Key::Right),
-        Keysym::Up => return Some(Key::Up),
-        Keysym::Down => return Some(Key::Down),
-        Keysym::Home => return Some(Key::Home),
-        Keysym::End => return Some(Key::End),
-        _ => {}
-    }
-
-    // Character input - use utf8 if available
-    if let Some(text) = utf8
-        && let Some(c) = text.chars().next()
-    {
-        // Only return printable characters or control characters we care about
-        if !c.is_control() || c == '\n' || c == '\r' || c == '\t' {
-            return Some(Key::Char(c));
-        }
-    }
-
-    // Fallback: convert keysym directly for release events where utf8 is always None.
-    // On press, utf8 = None means a compose sequence is in progress — don't insert anything.
-    if !is_press {
-        let raw = keysym.raw();
-
-        // Printable ASCII range (space through tilde): 0x20-0x7E
-        // XKB keysyms for these characters have the same value as ASCII
-        if (0x20..=0x7e).contains(&raw) {
-            return Some(Key::Char(char::from_u32(raw)?));
-        }
-
-        // Handle keypad numbers (KP_0 through KP_9)
-        // XKB_KEY_KP_0 = 0xffb0, XKB_KEY_KP_9 = 0xffb9
-        if (0xffb0..=0xffb9).contains(&raw) {
-            return Some(Key::Char(char::from_u32(raw - 0xffb0 + 0x30)?)); // Convert to '0'-'9'
-        }
-    }
-
-    None
 }
 
 impl SessionLockHandler for WaylandState {
@@ -2204,8 +1495,4 @@ impl ProvidesRegistryState for WaylandState {
 delegate_compositor!(WaylandState);
 delegate_output!(WaylandState);
 delegate_layer!(WaylandState);
-delegate_seat!(WaylandState);
-delegate_pointer!(WaylandState);
-delegate_touch!(WaylandState);
-delegate_keyboard!(WaylandState);
 delegate_registry!(WaylandState);
